+++ modules/twitter/twt-roles-list.go
@@ -105,7 +105,7 @@ func twtRoleListExec(ctx router.CommandCtx) {
 	numOfMentionsFooter := fmt.Sprintf(
 		"%d %s",
 		numOfMentions,
-		util.Pluralise("Mention Role", numOfMentions),
+		util.Pluralise("Mention Role", int64(numOfMentions)),
 	)
 
 	for i, description := range descriptionPages {
