+++ modules/vlive/vlive-feeds-list.go
@@ -135,7 +135,7 @@ func vliveFeedListExec(ctx router.CommandCtx) {
 	numOfFeedsFooter := fmt.Sprintf(
 		"%d %s",
 		numOfFeeds,
-		util.Pluralise("Feed", numOfFeeds),
+		util.Pluralise("Feed", int64(numOfFeeds)),
 	)
 
 	for i, description := range descriptionPages {
